more appeasing gcc

+++ ft_strwjoin.c
@@ -20,7 +20,7 @@ int	ft_strwjoin(char **old, const char *new, int len)
 	int	old_len;
 	int	total_len;
 
-	len = len < 0 ? ft_strlen(new) : len;
+	len = len < 0 ? (int)ft_strlen(new) : len;
 	old_len = *old == NULL
 		? 0
 		: ft_strlen(*old);
